Moves source files to src folder and renames them
